refactor: rename --> TelemetryFile
